docker_kafka/kafka-activity-tracker/producer/requirements.txt
flask
kafka-python

